Creates a regresion table in APA style with bootstrap confidence intervals (Beta version)

Creates a regresion table in APA style with bootstrap confidence intervals (Beta version)

apa.reg.boot.table(..., filename = NA, table.number = NA,
  number.samples = 1000)

Arguments

...
Regression (i.e., lm) result objects. Typically, one for each block in the regression.

filename
(optional) Output filename document filename (must end in .rtf or .doc only)

table.number
Integer to use in table number output line

number.samples
Number of samples to create for bootstrap CIs

conf.level
Confidence level for confidence intervals. Default is .95.

Value

APA table object

References

Algina, J., Keselman, H. J., & Penfield, R. J. (2008). Note on a Confidence Interval for the Squared Semipartial Correlation Coefficient. Educational and Psychological Measurement, 68(5), 734–741. http://doi.org/10.1177/0013164407313371

Examples

# View top few rows of goggles data set
# from Discovering Statistics Using R
head(album)

#>    adverts sales airplay attract
#> 1   10.256   330      43      10
#> 2  985.685   120      28       7
#> 3 1445.563   360      35       7
#> 4 1188.193   270      33       7
#> 5  574.513   220      44       5
#> 6  568.954   170      19       5

# Single block example
blk1 <- lm(sales ~ adverts + airplay, data=album)
apa.reg.boot.table(blk1,filename="exRegTable.doc")

#> Error in apa.reg.boot.table(blk1, filename = "exRegTable.doc"): could not find function "apa.reg.boot.table"

# Two block example, more than two blocks can be used
blk1 <- lm(sales ~ adverts, data=album)
blk2 <- lm(sales ~ adverts + airplay + attract, data=album)
apa.reg.boot.table(blk1,blk2,
                   filename="exRegBlocksTable.doc",
                   number.samples = 100) #number.samples > 1000 suggested

#> Error in apa.reg.boot.table(blk1, blk2, filename = "exRegBlocksTable.doc",     number.samples = 100): could not find function "apa.reg.boot.table"

# Interaction product-term test with blocks
blk1 <- lm(sales ~ adverts + airplay, data=album)
blk2 <- lm(sales ~ adverts + airplay + I(adverts * airplay), data=album)
apa.reg.boot.table(blk1,blk2,
                   filename="exInter1.doc",
                   number.samples=100) ##number.samples > 1000 suggested

#> Error in apa.reg.boot.table(blk1, blk2, filename = "exInter1.doc", number.samples = 100): could not find function "apa.reg.boot.table"

# Interaction product-term test with blocks and additional product terms
blk1<-lm(sales ~ adverts + airplay, data=album)
blk2<-lm(sales ~ adverts + airplay + I(adverts*adverts) + I(airplay*airplay), data=album)
blk3<-lm(sales~adverts+airplay+I(adverts*adverts)+I(airplay*airplay)+I(adverts*airplay),data=album)
apa.reg.boot.table(blk1,blk2,blk3,
                   filename="exInter2.doc",
                   number.samples=100) #number.samples > 1000 suggested

#> Error in apa.reg.boot.table(blk1, blk2, blk3, filename = "exInter2.doc",     number.samples = 100): could not find function "apa.reg.boot.table"

#Interaction product-term test with single regression (i.e., semi-partial correlation focus)
blk1 <- lm(sales ~ adverts + airplay + I(adverts * airplay), data=album)
apa.reg.boot.table(blk1,
                   filename="exInte3.doc",
                   number.samples=100) #number.samples > 1000 suggested

#> Error in apa.reg.boot.table(blk1, filename = "exInte3.doc", number.samples = 100): could not find function "apa.reg.boot.table"