Creates a table of means and standard deviations for a 1-way ANOVA design in APA style

Creates a table of means and standard deviations for a 1-way ANOVA design in APA style

apa.1way.table(iv, dv, data, filename = NA, table.number = NA,
  show.conf.interval = FALSE, landscape = FALSE)

Arguments

iv
Name of independent variable column in data frame

dv
Name of dependent variable column in data frame

data
Project data frame name

filename
(optional) Output filename document filename (must end in .rtf or .doc only)

table.number
Integer to use in table number output line

show.conf.interval
(TRUE/FALSE) Display confidence intervals in table.

landscape
(TRUE/FALSE) Make RTF file landscape

Value

APA table object

Examples

# Example 1: 1-way from Field et al. (2012) Discovery Statistics Using R
apa.1way.table(iv=dose,dv=libido,data=viagra,filename="ex1_desc_table.doc")

#> 
#> 
#> Descriptive statistics for libido as a function of dose.  
#> 
#>       dose    M   SD
#>  High Dose 5.00 1.58
#>   Low Dose 3.20 1.30
#>    Placebo 2.20 1.30
#> 
#> Note. M and SD represent mean and standard deviation, respectively.
#>  
#>