Creates a fixed-effects ANOVA table in APA style

Creates a fixed-effects ANOVA table in APA style

apa.aov.table(lm_output, filename, table.number = NA, conf.level = 0.9,
  type = 3)

Arguments

lm_output
Regression (i.e., lm) result objects. Typically, one for each block in the regression.

filename
(optional) Output filename document filename (must end in .rtf or .doc only)

table.number
Integer to use in table number output line

conf.level
Level of confidence for interval around partial eta-squared (.90 or .95). A value of .90 is the default, this helps to create consistency between the CI overlapping with zero and conclusions based on the p-value.

type
Sum of Squares Type. Type II or Type III; specify, 2 or 3, respectively. Default value is 3.

Value

APA table object

References

Smithson, M. (2001). Correct confidence intervals for various regression effect sizes and parameters: The importance of noncentral distributions in computing intervals. Educational and Psychological Measurement, 61(4), 605-632.

Fidler, F., & Thompson, B. (2001). Computing correct confidence intervals for ANOVA fixed-and random-effects effect sizes. Educational and Psychological Measurement, 61(4), 575-604.

Examples

#Example 1: 1-way from Field et al. (2012) Discovery Statistics Using R
options(contrasts = c("contr.helmert", "contr.poly"))
lm_output <- lm(libido ~ dose, data = viagra)
apa.aov.table(lm_output, filename = "ex1_anova_table.doc")

#> 
#> 
#> ANOVA results using libido as the dependent variable
#>  
#> 
#>    Predictor     SS df     MS     F    p partial_eta2 CI_90_partial_eta2
#>  (Intercept) 180.27  1 180.27 91.66 .000                                
#>         dose  20.13  2  10.06  5.12 .025          .46         [.04, .62]
#>        Error  23.60 12   1.97                                           
#> 
#> Note: Values in square brackets indicate the bounds of the 90% confidence interval for partial eta-squared 
#> 

# Example 2: 2-way from Fidler & Thompson (2001)
# You must set these contrasts to ensure values match SPSS
options(contrasts = c("contr.helmert", "contr.poly"))
lm_output <- lm(dv ~ a*b, data = fidler_thompson)
apa.aov.table(lm_output,filename = "ex2_anova_table.doc")

#> 
#> 
#> ANOVA results using dv as the dependent variable
#>  
#> 
#>    Predictor     SS df     MS      F    p partial_eta2 CI_90_partial_eta2
#>  (Intercept) 150.00  1 150.00 150.00 .000                                
#>            a   1.50  1   1.50   1.50 .238          .09         [.00, .32]
#>            b  12.00  3   4.00   4.00 .027          .43         [.04, .57]
#>        a x b   4.50  3   1.50   1.50 .253          .22         [.00, .38]
#>        Error  16.00 16   1.00                                            
#> 
#> Note: Values in square brackets indicate the bounds of the 90% confidence interval for partial eta-squared 
#> 

#Example 3: 2-way from Field et al. (2012) Discovery Statistics Using R
# You must set these contrasts to ensure values match SPSS
options(contrasts = c("contr.helmert", "contr.poly"))
lm_output <- lm(attractiveness ~ gender*alcohol, data = goggles)
apa.aov.table(lm_output, filename = "ex3_anova_table.doc")

#> 
#> 
#> ANOVA results using attractiveness as the dependent variable
#>  
#> 
#>         Predictor        SS df        MS       F    p partial_eta2
#>       (Intercept) 163333.33  1 163333.33 1967.03 .000             
#>            gender    168.75  1    168.75    2.03 .161          .05
#>           alcohol   3332.29  2   1666.14   20.07 .000          .49
#>  gender x alcohol   1978.12  2    989.06   11.91 .000          .36
#>             Error   3487.50 42     83.04                          
#>  CI_90_partial_eta2
#>                    
#>          [.00, .18]
#>          [.28, .60]
#>          [.15, .49]
#>                    
#> 
#> Note: Values in square brackets indicate the bounds of the 90% confidence interval for partial eta-squared 
#>